* Initial tests and fixes are complete * Most headers linked in GGEMain moved to source file to reduce compilation time * Unregister function for mouse Events * Move window function * When a pointer is added by Add function, if there is no base pointer the newly added pointer will be considered as base

+++ Resource/BitmapFontResource.h
@@ -40,7 +40,7 @@ namespace gge { namespace resource {
 		ImageResource *Characters[256];
 
 		////Horizontal separation distance between two characters
-		short Seperator;
+		int Seperator;
 		////Vertical spacing of this font (px)
 		int VerticalSpacing;
 		////Baseline from the top
